Much better, rebuild session fill

+++ src/main/java/com/taxah/diplomdb/repository/CheckRepository.java
@@ -3,5 +3,5 @@ package com.taxah.diplomdb.repository;
 import com.taxah.diplomdb.model.Check;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CheckRepository extends JpaRepository<Check, Integer>{
+public interface CheckRepository extends JpaRepository<Check, Long>{
 }
